+++ generic/tensor.h
@@ -6,13 +6,13 @@
 
 typedef struct THMKLTensor
 {
-	THTensor *tensor;   
-	long workspace[2]; 
-	char freeFlag;
-    int mklStorage;  //0:storage buffer allocated by THTensor, 1:storage buffer allocated by mklnn
-    long mkldnnLayout;
-    char flagBackup;
-    long * size;
+  THTensor *tensor;   
+  long workspace[2]; 
+  char freeFlag;
+  int mklStorage;  //0:storage buffer allocated by THTensor, 1:storage buffer allocated by mklnn
+  long mkldnnLayout;
+  char flagBackup;
+  size_t * size;
 } THMKLTensor;
 
 #define MKL_TENSOR_FLAG  32
@@ -21,7 +21,7 @@ void TH_MKL_(copyFromTH)(THMKLTensor * pTensor, THTensor * src);
 void TH_MKL_(copyBacktoTH)(THTensor * pTensor, THMKLTensor * src);
 void TH_MKL_(TH2MKL)(THMKLTensor * pTensor, THTensor * src);
 void TH_MKL_(MKL2TH)(THTensor * pTensor, THMKLTensor * src);
-void TH_MKL_(resize4d)(THMKLTensor *self, long size0, long size1, long size2, long size3);
+void TH_MKL_(resize4d)(THMKLTensor *self, size_t size0, size_t size1, size_t size2, size_t size3);
 void TH_MKL_(resizeAs)(THMKLTensor *self, THMKLTensor *src);
 real* TH_MKL_(data)(THMKLTensor *self);
 int TH_MKL_(nElement)(const THMKLTensor *self);
